GNG-480: Refactor server.js into separate routes files
