+++ routes/index.js
@@ -1,0 +1,6 @@
+const pageRouter = require("./page");
+const pagesRouter = require("./pages");
+const userRouter = require("./user");
+const usersRouter = require("./users");
+
+module.exports = { pageRouter, pagesRouter, userRouter, usersRouter };
